inference/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
numpy>=1.26.4
scipy>=1.14.0
torch>=2.2.1
torchvision>=0.17.1
python-dotenv>=1.0.1
gunicorn>=21.2.0
faststream>=0.5.25
PyWavelets>=1.4.1
tenacity>=8.2.0
prometheus_client>=0.20.0
psutil>=5.9.8
